naver-keyword-tool/requirements.txt
requests==2.32.3
pandas==2.2.3
python-dotenv==1.0.1
flask==3.1.0
google-ads==29.1.0
google-auth-oauthlib==1.2.1
